% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{an1}
\alias{an1}
\title{Sample Annotation Data}
\format{
## `an1`
A data frame with 644 rows and 9 columns:
\describe{
  \item{Criteria}{matching criteria for annotation}
  \item{Level}{confidence level}
  \item{Alignment.ID}{peak ID from LC-MS/MS}
  \item{Average.Mz}{peak m/z}
  \item{Metabolite.name}{compound name}
  \item{Adduct.type}{adduct}
  \item{Reference.mz}{reference library m/z}
  \item{S_N.average}{signal/noise ratio}
  \item{Source}{tissue source}
  ...
}
}
\source{
Stevens et. al. 2024
}
\usage{
an1
}
\description{
A subset of data from Stevens et. al. 2024 containing LC-MS/MS based
annotations for validating peak IDs from a mass spectrometry imaging
dataset.
}
\keyword{datasets}
