% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_norm}
\alias{d_norm}
\title{Demo Data}
\format{
## `d_norm`
A list of length 2:
\describe{
  \item{Sample.Info}{
A data frame containing input params. from msi_input_param()}
  \item{Data.files}{Two separate MSImagingExperiment objects
containing test data}
  ...
}
}
\source{
Stevens et. al. 2024
}
\usage{
d_norm
}
\description{
A subset of data from Stevens et. al. 2024 for two
samples acquired by MALDI-TOF MS processed by msi_peak_proc().
}
\keyword{datasets}
