% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md1}
\alias{md1}
\title{Sample Metadata Data}
\format{
## `md1`
A data frame with 2 rows and 4 columns:
\describe{
  \item{ID}{sample ID}
  \item{file.name}{file name and extension}
  \item{sample.name}{alternate sample name}
  \item{Group}{treatment group}
  ...
}
}
\source{
Stevens et. al. 2024
}
\usage{
md1
}
\description{
A subset of metadata from Stevens et. al. 2024 for two
samples acquired by MALDI-TOF MS.
}
\keyword{datasets}
