% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2c.process.peaks.annotation.R
\name{MSI.annotate.peaks}
\alias{MSI.annotate.peaks}
\title{Peak Annotation of MSI Data}
\usage{
MSI.annotate.peaks(pk.list, an.list)
}
\arguments{
\item{pk.list}{A data frame containing a data frame of processed peaks and processing details created by MSI.detectpeaks().}

\item{an.list}{A data frame containing annotated MSI peaks matched to a separate LC-MS/MS dataset.}
}
\value{
A filtered data frame and a table written to the disk containing annotated MSI peaks matched to a separate LC-MS/MS dataset.
}
\description{
Annotates binned peak lists by matching m/z values in complimentary LC-MS/MS data acquired
separately from an MSI experiment. Defaults to matching m/z values between MSI and LC-MS/MS data within 10 mDa.
}
\examples{
list.d <- MSI.LoadAll(list.p)
d.peaks <- MSI.binpeaks(list.d[["Sample.Info"]],list.d[["Data.files"]],d.peaks,TRUE)
list.anno <- read.table("annotations.neg.txt",sep = "\t",header = T)
MSI.annotate.peaks(d.peaks,list.anno)

}
