% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2b.process.peaks.binning.R
\name{MSI.binpeaks}
\alias{MSI.binpeaks}
\title{Parallel Peak Binning for MSI Datasets}
\usage{
MSI.binpeaks(samp.list, data.list, pk.list, parl)
}
\arguments{
\item{samp.list}{A list of Cardinal MSImagingExperiment sample information created from MSI.LoadAll().}

\item{data.list}{A list of Cardinal MSImagingExperiment objects created from MSI.LoadAll().}

\item{pk.list}{A list containing a data frame of processed peaks and processing details created by MSI.detectpeaks().}

\item{parl}{Should processing be run in parallel? Note: only supported by WSL2 or Linux; will default to sequential processing if toggled 'TRUE' on Windows OS.}
}
\value{
A list containing a data frame of processed peaks and processing details.
}
\description{
Converts profile spectra to centroided data based on a peak list generated by MSI.detectpeaks().
Runs in parallel using 50% of available cores if run on WSL2 or Linux, and splits binning into batches
of 3 samples or less to conserve resources. For use with older versions of Cardinal (< 3.6.3).
}
\examples{
MSI.binpeaks(list.d[["Sample.Info"]],list.d[["Data.files"]],d.peaks,TRUE)

}
