% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1a.stats.functions.R
\name{MSI.ChemRICH.input}
\alias{MSI.ChemRICH.input}
\title{Merge Function for ChemRICH Analysis}
\usage{
MSI.ChemRICH.input(pval.df, fold.df, feat.df, g.var)
}
\arguments{
\item{pval.df}{One-way ANOVA results data frame.}

\item{fold.df}{Group-wise fold change data frame.}

\item{feat.df}{Data frame containing feature metadata.}

\item{g.var}{Treatment group column name.}
}
\value{
A merged data frame formatted for enrichment analysis based on user-defined compound classes.
}
\description{
Combines one-way ANOVA and fold change results for enrichment analysis based on user-defined compound classes.
This enrichment analysis method is adapted from Chemical Similarity Enrichment Analysis (ChemRICH), initially described
in Barupal et. al. 2017 (DOI: 10.1038/s41598-017-15231-w).
}
\examples{
# cr.in <- MSI.ChemRICH.input(
# # p-value input df
# d.p2.result,
# # fold change input df
# d.fold,
# # feature metadata
# md.feat,
# # grouping column for calculating ChemRICH
# "label.saturation"
# )

}
