% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2c.process.peaks.annotation.R
\name{MSI.data.filter}
\alias{MSI.data.filter}
\title{Peak Annotation of Binned MSI Data}
\usage{
MSI.data.filter(proc.data, an.list.filt)
}
\arguments{
\item{proc.data}{Processed MSImagingExperiment object.}

\item{an.list.filt}{A filtered data frame containing annotated MSI peaks from MSI.annotate.peaks().}
}
\value{
A combined Cardinal MSImagingExperiment containing annotated features for all study samples.
}
\description{
Annotates binned MSI data and filters peaks for removing duplicate matched m/z values in complimentary LC-MS/MS data acquired
separately from an MSI experiment.
}
\examples{
d.anno <- MSI.data.filter(d3,list.annotated[[1]])

### Verify that all samples have been correctly annotated
Cardinal::imageData(d.anno)
Cardinal::features(d.anno)
Cardinal::pixelData(d.anno)
Cardinal::run(d.anno)
Cardinal::spectra(
  d.anno,
  "intensity"
  )[,1:25]

### Save data
saveRDS(
  d.anno,
  "data.processed.rds"
  )
gc(reset = T)

}
