% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3c.normalize.check.images.R
\name{MSI.image.check}
\alias{MSI.image.check}
\title{Ion images for comparing normalization performance}
\usage{
MSI.image.check(df, md.df, var.int, var.g, perc.int)
}
\arguments{
\item{df}{A data matrix containing grouping information and normalized pixel intensities for each annotated compound.}

\item{md.df}{A data matrix containing pixel metadata for each annotated compound.}

\item{var.int}{Variable corresponding to the pixel intensity for each sample.}

\item{var.g}{Sample ID variable.}

\item{perc.int}{quantile to be used for the intensity color scale to improve contrast in ion images.}
}
\value{
An ion image visualizing pixel intensities to compare normalization performance.
}
\description{
Visualizes a series of ion images comparing various normalization methods for each imaging run.
}
\examples{
# Generate ion images for all annotated compounds
parallel::mclapply(
  mc.cores = ceiling(
    parallel::detectCores()*
      0.9
  ),
  seq.int(
    6,
    (ncol(l.norm[["norm.sLOESS"]]) + 2),
    1
  ),
  function(x)
    ggplot2::ggsave(
      paste(
        "plot.check.",
        names(l.norm[["norm.sLOESS"]][x - 2]),
        ".png",
        sep = ""),
      MSI.image.check(
        # Input data
        l.norm[["norm.sLOESS"]],
        # Input metadata
        msi.norm.form[["Pixel"]],
        # Intensity to plot
        x,
        # Sample ID variable
        "ID",
        # Percentile for scaling contrast of images
        0.99
      ),
      width = 15,
      height = 6,
      dpi = 1200
    )
)

}
