% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.data.import.R
\name{MSI.InputParams}
\alias{MSI.InputParams}
\title{Input Parameters for Loading MSI Datasets in .imzML Format}
\usage{
MSI.InputParams(dp, pol, mz.low, mz.high, res)
}
\arguments{
\item{dp}{data folder location, relative to R project directory}

\item{pol}{ionization mode, either 'pos' or 'neg'. The ionization mode must be present in the corresponding file names.}

\item{mz.low}{lower limit of mass range in units of m/z}

\item{mz.high}{upper limit of mass range in units of m/z}

\item{res}{mass resolution, in ppm}
}
\value{
A list containing .imzML input parameters and file names.
}
\description{
Specifies input parameters to be read into R by the Cardinal package.
The user must provide the path to paired .imzML and .ibd files, if applicable,
for subsequent loading as a single list of MSImagingExperiment objects created
by Cardinal.
}
\examples{
MSI.InputParams("data/","neg",300,900,5)

}
