% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.data.import.R
\name{MSI.LoadAll}
\alias{MSI.LoadAll}
\title{Load MSI Datasets in .imzML Format}
\usage{
MSI.LoadAll(list.p)
}
\arguments{
\item{list.p}{list containing input parameters and file names created from MSI.InputParams()}
}
\value{
A list of Cardinal MSImagingExperiment objects created for each file included in the specified data folder.
}
\description{
Uses input parameters from MSI.InputParams() to read files into R using the Cardinal function readMSIData().
}
\examples{
MSI.LoadAll(list.p)

}
