% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{MSI.markers.top5}
\alias{MSI.markers.top5}
\title{Top-5 Cluster Heatmap}
\usage{
MSI.markers.top5(so, marks, var.clus, h.w, h.h, fs.c, fs.r, samp.no)
}
\arguments{
\item{so}{Seurat object containing segmented MSI data.}

\item{marks}{List of cluster marker compounds to use for subsetting Seurat object.}

\item{var.clus}{Cluster variable.}

\item{fs.c}{Heatmap column font size.}

\item{fs.r}{Heatmap row font size.}

\item{hm.w}{Heatmap width.}

\item{hm.h}{Heatmap height.}

\item{sample.no}{Sample ID number.}
}
\value{
A heatmap displaying the top-5 markers for each cluster.
}
\description{
Plots a heatmap of the top-5 compounds represented within each segmented cluster in a MSI Seurat object.
}
\examples{
## Change to lapply if not running on Linux or WSL2
d.seg.mark <- setNames(
parallel::mclapply(
  mc.cores = ceiling(
    parallel::detectCores()*
      0.5
  ),
  seq.int(1,length(d.seg),1),
  function(x){
    # Find marker compounds for each calculated cluster
    d.mark <- Seurat::FindAllMarkers(
      d.seg[[x]],
      test.use = "wilcox",
      densify = T,
      verbose = T
    )
    names(d.mark) <- c(
      names(
        d.mark[
          ,
          c(1:6)]
      ),
      "name"
    )
    d.mark[["ID"]] <- names(d.seg[x])

    # Create top-5 heatmap
    h.out <- MSI.markers.top5(
      # Seurat object
      d.seg[[x]],
      # Marker gene list
      d.mark,
      # Cluster column
      "Cluster",
      # Heatmap width
      24,
      # Heatmap height
      12,
      # Column fontsize
      6,
      # Row fontsize
      8,
      # Sample number
      x
    )

    return(
      list(
        "Markers" = d.mark,
        "Plot" = h.out
      )
    )
  }
),
c(names(d.seg))
)

}
