% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3a.normalize.format.R
\name{MSI.norm.format}
\alias{MSI.norm.format}
\title{Format processed MSI data for normalization}
\usage{
MSI.norm.format(msi.d, an.list, p.md)
}
\arguments{
\item{msi.d}{A dense matrix containing MSI data from all annotated compounds in each study sample.}

\item{an.list}{A data frame containing annotated MSI peaks matched to a separate LC-MS/MS dataset.}

\item{p.md}{A data frame containing treatment and group information for each imaging run.}
}
\value{
Formatted pixel, feature, and image data for downstream normalization and statistical analysis.
}
\description{
Annotates binned peak lists by matching m/z values in complimentary LC-MS/MS data acquired
separately from an MSI experiment. Defaults to matching m/z values between MSI and LC-MS/MS data within 10 mDa.
}
\examples{
# Load processed data, annotation list, and sample metadata table
## Processed data
if(
  !exists("d") &
  file.exists("data.processed.rds")
  )
 {
  d <- readRDS(
    "data.processed.rds"
    )
 }
Cardinal::imageData(d)

## Annotation list
d.anno <- read.table(
  "data.annotated.neg.curated.txt",
  header = T,
  sep = "\t"
)

## pixel metadata
d.group <- read.table(
  "sample.metadata.neg.txt",
  sep = "\t",
  header = T
)

## Format data (also saves formatted outputs)
msi.norm.form <- MSI.norm.format(
  # Image data
  d,
  # Curated annotation list
  d.anno,
  # Pixel metadata
  d.group
)

}
