% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{MSI.norm.sLOESS}
\alias{MSI.norm.sLOESS}
\title{Normalization methods for MSI data}
\usage{
MSI.norm.sLOESS(msi.mat, f, p, parl, core.perc)
}
\arguments{
\item{msi.mat}{A dense matrix containing MSI data from all annotated compounds in each study sample.}

\item{f}{A data frame containing metadata for each annotated compound.}

\item{p}{A data frame containing formatted pixel metadata for each imaging run.}

\item{parl}{Logical indicating whether data normalization for each compound should be run in parallel (Linux and WSL2 only).}

\item{core.perc}{Percentage of available cores to use if running in parallel (Linux and WSL2 only).}
}
\value{
A data matrix containing grouping information and normalized pixel intensities for each annotated compound.
}
\description{
Various normalization methods for MSI data. MSI.norm.none() only calculates the TIC of each pixel
but performs no normalization. This method is useful for comparing overall normalization performance.
MSI.norm.TIC() performs total ion current normalization for each annotated compound. MSI.norm.sLOESS()
implements sparse LOESS, which excludes pixels with zero signal intensity to preserve biological differences
in compound spatial distribution while removing systematic artifacts incurred during data acquisition.
}
\examples{
# Normalizations
l.norm <- list(
  "norm.none" = MSI.norm.none(
    msi.norm.form[["Data"]],
    msi.norm.form[["Feature"]],
    msi.norm.form[["Pixel"]]
  ),
  "norm.TIC" = MSI.norm.TIC(
    msi.norm.form[["Data"]],
    msi.norm.form[["Feature"]],
    msi.norm.form[["Pixel"]]
  ),
  "norm.sLOESS" = MSI.norm.sLOESS(
    # Data matrix
    msi.norm.form[["Data"]],
    # Feature metadata
    msi.norm.form[["Feature"]],
    # Pixel metadata
    msi.norm.form[["Pixel"]],
    # Process in parallel (TRUE/FALSE)
    TRUE,
    # Percent of total cores to use (as decimal)
    0.9
  )
)

}
