% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{MSI.plot.AllClusters}
\alias{MSI.plot.AllClusters}
\title{Composite cluster images for all samples}
\usage{
MSI.plot.AllClusters(no, clus1)
}
\arguments{
\item{no}{Numeric sample ID.}

\item{clus1}{Name of clustering variable (most often this is 'Cluster').}
}
\value{
A panel of ion images stratified by sample ID for establishing distinct morphological regions in each sample.
}
\description{
Visualizes combined cluster images for all samples. The input of this function requires a pre-existing normalized data matrix from
the previous steps, which is merged with the corresponding cluster column from the segmented Seurat object and indexed by sample ID and
pixel number.
}
\examples{
# Add Seurat clustering column to image data frame and overwrite
d <- dplyr::left_join(
  dplyr::bind_rows(
    lapply(
      seq.int(1,length(d.seg),1),
      function(x)
        data.frame(
          "pixel" = d[
            d[["ID"]] == as.numeric(names(d.seg[x])),
            "pixel"
          ],
          "Cluster" = as.factor(as.numeric(
            d.seg[[x]]@meta.data[["seurat_clusters"]]
          )
          ))
    )
  ),
  d,
  by = "pixel"
)

d <- dplyr::left_join(
  msi.pmd[
    ,
    c("pixel","X","Y")
  ],
  d,
  by = "pixel"
)

# Plot cluster images for each sample

d.seg.img <- setNames(
  lapply(
    seq.int(
      1,
      length(
        unique(
          d[,c("ID")]
        )
      ),
      1
    ),
    function(x){
      MSI.plot.AllClusters(
        # input data
        x,
        # cluster column
        "Cluster"
      )
    }
  ),
  c(
    unique(
      d[
        ,
        c("ID")
      ]
    )
  )
)

}
