% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{MSI.plot.AllImage}
\alias{MSI.plot.AllImage}
\title{Ion images for comparing normalization performance}
\usage{
MSI.plot.AllImage(df, var.int, var.g, perc.int)
}
\arguments{
\item{df}{A data matrix containing grouping information and normalized pixel intensities for each annotated compound.}

\item{var.int}{Variable corresponding to the pixel intensity for each sample.}

\item{var.g}{Sample ID variable.}

\item{perc.int}{quantile to be used for the intensity color scale to improve contrast in ion images.}
}
\value{
An ion image visualizing pixel intensities to compare normalization performance.
}
\description{
Visualizes a series of ion images comparing various normalization methods for each imaging run.
}
\examples{
## Plot TIC of each normalization for each sample
ggplot2::ggsave(
  "plot.norm.image.sLOESS.png",
  MSI.plot.AllImage(
    # Input df
    msi.norm.form[["Pixel"]],
    # Intensity to plot
    "mTIC.pixel.sLOESS",
    # Sample ID variable
    "ID",
    # Percentile for scaling contrast of images
    0.9
  ),
  width = 6,
  height = 10,
  dpi = 300
)

}
