% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.chemrich.R
\name{MSI.plot.ChemRICH}
\alias{MSI.plot.ChemRICH}
\title{Plot ChemRICH Results}
\usage{
MSI.plot.ChemRICH(df, xlab1, p.title)
}
\arguments{
\item{df}{ChemRICH results data frame.}

\item{xlab1}{X-axis label name.}

\item{p.title}{Plot title.}
}
\value{
A dot plot displaying ChemRICH results, including cluster size, increased/decreased proportion, and an indication of statistical significance.
}
\description{
Plots ChemRICH results for selected comparison as a dot plot.
}
\examples{
chemrich.plot <- MSI.run.ChemRICH(
chemrich.input,
7,
16,
3,
"Class"
)

}
