% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.3.segment.marker.vis.R
\name{MSI.plot.coloc}
\alias{MSI.plot.coloc}
\title{Cluster/Region Colocalization}
\usage{
MSI.plot.coloc(df, md, samp.no, list.cols, name.region, region.clus.no, no.c)
}
\arguments{
\item{df}{Normalized data matrix containing cluster, sample ID, pixel coordinate, and pixel intensity information.}

\item{md}{Numeric sample ID.}

\item{samp.no}{Numeric sample ID.}

\item{list.cols}{Numeric IDs indicating the column numbers of compounds to be visualized.}

\item{name.region}{User supplied name of morphological region represented by the chosen cluster ID}

\item{region.clus.no}{Numeric cluster ID.}

\item{no.c}{Number of palette colors to be used (equal to the number of compound ion images to be plotted).}
}
\value{
A panel of ion images with cluster ID overlays for examining compound colocalization within distinct morphological regions in each sample.
}
\description{
Visualizes the signal intensity distribution of a selected compound with an overlay indicating a specified cluster obtained from
spatial segmentation results. Some underlying information is required, such as the morphological region represented by the chosen
cluster. This function is designed to accept a range of inputs but relies on user-based curation and region assignment of segmentation
results.
}
\examples{
# Create plots and save (requires input data frame with segmented clusters from previous step)
p.mark <- MSI.plot.coloc(
  # input data
  d,
  # number of metadata columns using for input
  6,
  # sample (ID) number
  1,
  # list of input data columns to include (by number)
  c(
    2:5,49:50,53,55
  ),
  # user-defined region name
  "Alveolar Epithelium",
  # cluster number of selected region (for selected sample)
  3,
  # number of colors from palette to use
  5
)

}
