% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{MSI.plot.IndClusters}
\alias{MSI.plot.IndClusters}
\title{Individual cluster images for each sample}
\usage{
MSI.plot.IndClusters(df, no.s, no.c)
}
\arguments{
\item{df}{Normalized data matrix containing cluster, sample ID, and pixel coordinate information.}

\item{no.s}{Numeric sample ID.}

\item{no.c}{Numeric cluster ID.}
}
\value{
A panel of ion images stratified by cluster ID for establishing distinct morphological regions in each sample.
}
\description{
Visualizes split cluster images for an individual sample. The input of this function requires a pre-existing normalized data matrix from
the previous steps, which is merged with the corresponding cluster column from the segmented Seurat object and indexed by sample ID and
pixel number.
}
\examples{
# Split composite cluster images into individual clusters for validation
d.seg.img.split <- setNames(
  lapply(
    seq.int(
      1,
      length(unique(d[,c("ID")])),
      1
    ),
    function(x) {
      setNames(
        lapply(
          seq.int(
            1,
            length(
              unique(
                d[
                  d[["ID"]] == x,
                ]
                [["Cluster"]]
              )
            ),
            1
          ),
          function(y)
            MSI.plot.IndClusters(
              # Input data
              d,
              # Sample number
              x,
              # Cluster number
              y
            )
        ),
        c(
          sort(
            unique(
              d[
                d[["ID"]] == x,
              ]
              [["Cluster"]]
            )
          )
        )
      )
    }
  ),
  c(unique(
    d[
      ,
      c("ID")
    ]
  )
  )
)

}
