% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{MSI.plot.OneImage}
\alias{MSI.plot.OneImage}
\title{Ion image for comparing normalization performance}
\usage{
MSI.plot.OneImage(df, var.int, var.g, samp.no, perc.int)
}
\arguments{
\item{df}{A data matrix containing grouping information and normalized pixel intensities for each annotated compound.}

\item{var.int}{Variable corresponding to the pixel intensity for each sample.}

\item{var.g}{Sample ID variable.}

\item{samp.no}{Specific sample ID to plot.}

\item{perc.int}{quantile to be used for the intensity color scale to improve contrast in ion images.}
}
\value{
An ion image visualizing pixel intensities to compare normalization performance.
}
\description{
Visualizes a single ion image comparing various normalization methods.
}
\examples{
ggplot2::ggsave(
  "plot.norm.image.compare.png",
  ggpubr::ggarrange(
    MSI.plot.OneImage(
      # Input df
      msi.norm.form[["Pixel"]],
      # Intensity to plot
      "mTIC.pixel.raw",
      # Sample ID variable
      "ID",
      # Sample number
      1,
      # Percentile for scaling contrast of images
      0.99
    ),
    MSI.plot.OneImage(
      # Input df
      msi.norm.form[["Pixel"]],
      # Intensity to plot
      "mTIC.pixel.TIC",
      # Sample ID variable
      "ID",
      # Sample number
      1,
      # Percentile for scaling contrast of images
      0.99
    ),
    MSI.plot.OneImage(
      # Input df
      msi.norm.form[["Pixel"]],
      # Intensity to plot
      "mTIC.pixel.sLOESS",
      # Sample ID variable
      "ID",
      # Sample number
      1,
      # Percentile for scaling contrast of images
      0.99
    ),
    labels = c(
      "Raw",
      "mTIC",
      "sLOESS"
    ),
    nrow = 1,
    common.legend = T,
    legend = "bottom"
  ),
  width = 24,
  height = 12,
  dpi = 600
)

}
