% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{MSI.plot.TIC}
\alias{MSI.plot.TIC}
\title{Scatterplot for comparing normalization performance}
\usage{
MSI.plot.TIC(df, var.x, var.d, var.g)
}
\arguments{
\item{df}{A data matrix containing grouping information and normalized pixel intensities for each annotated compound.}

\item{var.x}{Variable to be plotted on the x-axis (this is usually the acquisition order of each pixel).}

\item{var.d}{Variable to be plotted on the y-axis (this is usually the TIC intensity for each pixel).}

\item{var.g}{Variable to be used for grouping of scatterplot points (this is usually the imaging run ID for each sample).}
}
\value{
A scatter plot visualizing pixel intensities over time to compare normalization performance.
}
\description{
Visualizes a series of scatter plots comparing various normalization methods for MSI data.
}
\examples{
## Append Raw/TIC/sLOESS counts to sample metadata
msi.norm.form[["Pixel"]] <- data.frame(
  msi.norm.form[["Pixel"]],
  "mTIC.pixel.raw" = apply(
    l.norm[["norm.none"]][
      ,
      4:ncol(l.norm[["norm.none"]])
    ],
    1,
    function(x)
      sum(x)
  ),
  "mTIC.pixel.TIC" = l.norm[["norm.TIC"]][["mTIC.norm"]],
  "mTIC.pixel.sLOESS" = apply(
    l.norm[["norm.sLOESS"]][
      ,
      4:ncol(l.norm[["norm.sLOESS"]])
    ],
    1,
    function(x)
      sum(x)
  )
)

## Save plot
ggplot2::ggsave(
  "plot.norm.TIC.png",
  ggpubr::ggarrange(
    ## Raw
    MSI.plot.TIC(
      ## sample metadata
      msi.norm.form[["Pixel"]],
      ## pixel column
      "pixel",
      ## TIC for selected norm
      "mTIC.pixel.raw",
      ## Sample ID column
      "ID"
    ),
    MSI.plot.TIC(
      msi.norm.form[["Pixel"]],
      "pixel",
      "mTIC.pixel.TIC",
      "ID"
    ),
    MSI.plot.TIC(
      msi.norm.form[["Pixel"]],
      "pixel",
      "mTIC.pixel.sLOESS",
      "ID"
    ),
    ncol = 3,
    common.legend = T,
    labels = c("Raw","TIC","sLOESS")
  ),
  width = 18,
  height = 6,
  dpi = 600
)

}
