% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1b.stats.volcano.plot.R
\name{MSI.plot.vol}
\alias{MSI.plot.vol}
\title{Volcano Plot Function for Univariate Statistical Analysis}
\usage{
MSI.plot.vol(df, group_fc, group_p, name.lab, fc.lim, title)
}
\arguments{
\item{df}{Input data frame containing p-values and group fold changes for each compound.}

\item{group_fc}{Fold change variable name (provided as a character string).}

\item{group_p}{P-value variable name (provided as a character string).}

\item{name.lab}{Vector containing compound names.}

\item{fc.lim}{Numeric limit of the x-axis.}

\item{title}{Plot title.}
}
\value{
A volcano plot displaying significantly altered compounds comparing a specific treatment.
}
\description{
Uses the EnhancedVolcano package to create a volcano plot from a selection of p-values, fold changes, and compound names.
}
\examples{
## Volcano plot
ggsave(
  "plotname.png",
  MSI.plot.vol(
    data.frame(
      "Name" = names(d.p2.result[,2:ncol(d.p2.result)]),
      "FC" = t(d.fold[1,2:ncol(d.fold)])[,1],
      "P" = t(d.p2.result[2,2:ncol(d.p2.result)])[,1]
    ),
    "FC","P","Name",4,
    "Group name"
  ),
  width = 8,
  height = 6,
  dpi = 800
)

}
