% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2a.process.peaks.detection.R
\name{MSI.processpeaks}
\alias{MSI.processpeaks}
\title{Combined Peak Processing for MSI Datasets}
\usage{
MSI.processpeaks(msi.list, p.list, parl, core.perc)
}
\arguments{
\item{msi.list}{A list of Cardinal MSImagingExperiment objects created from MSI.LoadAll().}

\item{p.list}{A list containing input parameters and file names created from MSI.InputParams()}

\item{parl}{Should processing be run in parallel? Note: only supported by WSL2 or Linux; will default to sequential processing if toggled 'TRUE' on Windows OS.}

\item{core.perc}{Percentage of cores to be used during processing.}
}
\value{
A list containing a data frame of processed peaks and processing details.
}
\description{
Performs peak picking and alignment based on all data files included in MSI.LoadAll() using Cardinal.
}
\examples{
MSI.processpeaks(list.d[["Data.files"]],list.p,TRUE,0.75)

}
