% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{MSI.Segment}
\alias{MSI.Segment}
\title{Seurat-based KNN clustering and spatial segmentation}
\usage{
MSI.Segment(sample.no, md, span, clus.res)
}
\arguments{
\item{sample.no}{Sample ID number.}

\item{md}{Number of metadata columns present in data matrix.}

\item{span}{loess span parameter (proportion of pixels to fit model to; use lower proportion for tissues with small regions).}

\item{clus.res}{Resolution to use for clustering (see Seurat FindClusters() documentation for details; larger number = more clusters).}
}
\value{
A Seurat object containing segmented clusters for a specified sample.
}
\description{
Converts MSI data to a Seurat object and performs spatial segmentation using the default implementation of the KNN algorithm used by Seurat.
}
\examples{
# Find clusters for each sample
d.seg <- setNames(
  parallel::mclapply(
    mc.cores = ceiling(
      parallel::detectCores()*
        0.5
    ),
    seq.int(
      1,
      length(
        unique(
          d[,c("ID")]
        )
      ),
      1
    ),
    function(x)
      MSI.Segment(
        # sample number
        x,
        # metadata column number
        8,
        # loess span parameter (proportion of pixels to fit model to; use lower proportion for tissues with small regions)
        0.1,
        # resolution to use for clustering (see Seurat FindClusters() documentation for details; larger number, more clusters)
        0.9
      )
  ),
  c(
    unique(
      d[
        ,
        c("ID")
      ]
    )
  )
)

}
