% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1a.stats.functions.R
\name{MSI.stat.FC}
\alias{MSI.stat.FC}
\title{Group-wise Fold Change Function}
\usage{
MSI.stat.FC(df, md, grp)
}
\arguments{
\item{df}{Normalized data matrix containing cluster, sample ID, pixel coordinate, and pixel intensity information.}

\item{md}{Number of metadata columns present in input data.}

\item{grp}{Treatment group column name.}
}
\value{
A data frame containing the group-wise fold changes for each compound.
}
\description{
Calculates the log2 fold changes for all possible combinations of study treatment groups.
}
\examples{
# Calculate group-wise fold changes
l.fold <- setNames(
  lapply(
    seq.int(
      1,
      length(
        unique(
          d[["ID"]]
        )
      ),
      1
    ),
    function(f)
      MSI.stat.FC(
        # input data
        d[d[["ID"]] == f,],
        # metadata column number
        8,
        # cluster number column
        "Cluster"
      )
  ),
  c(
    unique(
      d[["ID"]]
    )
  )
)

}
