% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1a.stats.functions.R
\name{MSI.stat.output}
\alias{MSI.stat.output}
\title{Merge Function for one-way ANOVA and Fold Change Results}
\usage{
MSI.stat.output(pval.df, fold.df, g.var)
}
\arguments{
\item{pval.df}{One-way ANOVA results data frame.}

\item{fold.df}{Group-wise fold change data frame.}

\item{g.var}{Treatment group column name.}
}
\value{
A merged data frame containing one-way ANOVA and fold change results for each compound.
}
\description{
Combines one-way ANOVA and fold change results for selecting regional comparisons and visualizing results.
}
\examples{
# Combine results and select region comparisons for visualization
l.comb <- setNames(
  lapply(
    seq.int(
      1,
      length(
        unique(
          d[["ID"]]
        )
      ),
      1
    ),
    function(f)
      MSI.stat.output(
        # Statistics results df
        l.p[[f]][["FDR P"]][-1],
        # Fold change df
        l.fold[[f]],
        # Grouping variable
        "Comparison"
      )
  ),
  c(
    unique(
      d[["ID"]]
    )
  )
)

}
