% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1a.stats.functions.R
\name{MSI.stat.owANOVA}
\alias{MSI.stat.owANOVA}
\title{One-way ANOVA of MSI Data}
\usage{
MSI.stat.owANOVA(d, md, g)
}
\arguments{
\item{d}{Normalized data matrix containing cluster, sample ID, pixel coordinate, and pixel intensity information.}

\item{md}{Number of metadata columns present in input data.}

\item{g}{Treatment group column name.}
}
\value{
A data frame containing one-way ANOVA results for each compound.
}
\description{
Conducts a one-way ANOVA with Tukey's post-hoc analysis for each compound.
}
\examples{
# Calculate one-way ANOVA
l.p <- setNames(
  lapply(
    seq.int(
      1,
      length(
        unique(
          d[["ID"]]
        )
      ),
      1
    ),
    function(f)
      MSI.stat.owANOVA(
        # input data
        d[d[["ID"]] == f,],
        # metadata column number
        8,
        # cluster number column
        "Cluster"
      )
  ),
  c(
    unique(
      d[["ID"]]
    )
  )
)

}
