% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1a.stats.functions.R
\name{MSI.stat.pixel}
\alias{MSI.stat.pixel}
\title{Cluster Pixel Counts}
\usage{
MSI.stat.pixel(df, md, g.var1, g.var2)
}
\arguments{
\item{df}{Normalized data matrix containing cluster, sample ID, pixel coordinate, and pixel intensity information.}

\item{md}{Number of metadata columns present in input data.}

\item{g.var1}{Numeric sample ID column name.}

\item{g.var2}{Numeric cluster ID column name.}
}
\value{
A data frame containing the number of pixels per cluster for each sample.
}
\description{
Counts the number of pixels in each cluster for each sample.
}
\examples{
# Count pixel number per cluster for each sample
d.pix <- MSI.stat.pixel(
  # input data
  d,
  # metadata column number
  8,
  # sample number column
  "ID",
  # cluster number column
  "Cluster"
)

}
