% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.stats.univar.R
\name{msi_cnt_pix}
\alias{msi_cnt_pix}
\title{Count Pixels}
\usage{
msi_cnt_pix(mdf, var_sel)
}
\arguments{
\item{mdf}{A data frame containing MSI experiment metadata.}

\item{var_sel}{Names of metadata variables to count.}
}
\value{
A data frame containing the number of pixels per each
group specified in var_sel.
}
\description{
Counts the pixel number per each specified variable in a MSI
dataset. Can select up to 3 different variables for counting.
}
\examples{

# cnt_pix <- msi_cnt_pix(
#   mdf = d_seg[["pixels"]],
#   var_sel = c("Group", "ID", "Cluster")
# )

}
