% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.stats.univar.R
\name{msi_data_check}
\alias{msi_data_check}
\title{Data Transformation and Quality Check}
\usage{
msi_data_check(ld = d_seg[["data"]])
}
\arguments{
\item{ld}{A MSI data frame.}
}
\value{
List containing formatted input data and metadata
for downstream analysis.
}
\description{
Performs data imputation, log2-transformation, and
scaling of a MSI dataset. The default option is to use
segmented MSI data, stored as a variable named "d_seg."
The defaults can be overridden by specifying additional
function parameters. See ?RegioMSI::msi_data_check for
more details.
}
\examples{

# d_seg[["scale.data"]] <- msi_data_check()

## Specify data to use (replace with formatted data frames)
# d_seg[["scale.data"]] <- msi_data_check(ld = d_seg[["data"]])

}
