% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{msi_data_norm}
\alias{msi_data_norm}
\title{MSI Data Normalization}
\usage{
msi_data_norm(dm, feat, pix, mtd, parl = FALSE, core_perc = NULL)
}
\arguments{
\item{dm}{A MSI data matrix.}

\item{feat}{A data frame containing metadata for each annotated compound.}

\item{pix}{A data frame containing sample metadata for each pixel}

\item{mtd}{Normalization method (either "none", "tic", "rms", or "sLOESS").}

\item{parl}{(optional) Logical indicating if normalization
should be performed in parallel.}

\item{core_perc}{(optional) Percentage of cores to use for parallel
normalization.}
}
\value{
A list containing normalized pixel intensities and metadata.
}
\description{
Includes multiple methods for MSI data normalization. RegioMSI introduces
sparse locally estimated scatterplot smoothing (sLOESS) normalization, which
excludes pixels with zero signal intensity to preserve biological differences
in compound spatial distribution while removing systematic artifacts
incurred during data acquisition.
}
\examples{

# d_norm <- msi_data_norm(
#   dm = d_norm[["Data"]],
#   feat = d_norm[["Feature"]],
#   pix = d_norm[["Pixel"]],
#   mtd = "sLOESS",
#   parl = TRUE,
#   core_perc = 0.2
# )

# d_norm <- msi_data_norm(
#   dm = d_norm[["Data"]],
#   feat = d_norm[["Feature"]],
#   pix = d_norm[["Pixel"]],
#   mtd = "tic"
# )

}
