% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{msi_data_qc}
\alias{msi_data_qc}
\title{MSI Processing QC}
\usage{
msi_data_qc(dm, feat, pix, parl = FALSE, core_perc = NULL)
}
\arguments{
\item{dm}{A MSI data matrix.}

\item{feat}{A data frame containing metadata for each annotated compound.}

\item{pix}{A data frame containing sample metadata for each pixel}

\item{parl}{(optional) Logical indicating if normalization
should be performed in parallel.}

\item{core_perc}{(optional) Percentage of cores to use for parallel
normalization.}

\item{mtd}{Normalization method (either "none", "tic", "rms", or "sLOESS").}
}
\value{
A list containing normalized pixel intensities and metadata.
}
\description{
Calculates descriptive statistics for each feature in a formatted
MSI dataset. Can be used for subsequent artifact detection and
removal.
}
\examples{

# data_qc <- msi_data_qc(
#   dm = d_norm[["Data"]],
#   feat = d_norm[["Feature"]],
#   pix = d_norm[["Pixel"]],
#   parl = FALSE,
#   core_perc = 0.25
# )

}
