% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.data.import.R
\name{msi_input_param}
\alias{msi_input_param}
\title{Input Parameters for Loading MSI Datasets in .imzML Format}
\usage{
msi_input_param(dp, pol, mz_range, mz_res)
}
\arguments{
\item{dp}{Data path relative to the "data/" folder. Data files
should be separated by subfolders according to ionization mode.}

\item{pol}{Ionization mode (either "pos" or "neg").}

\item{mz_range}{A vector specifying the limits of the m/z range.}

\item{mz_res}{Mass resolution, in ppm.}
}
\value{
A data frame containing .imzML input parameters and file names.
}
\description{
Specifies input parameters to be read into R by the Cardinal package.
The user must provide the path to a folder containing .imzml files,
for creating a single list of MSImagingExperiment objects using
Cardinal. See doi:10.1007/978-1-60761-987-1_12 for more information
about .imzml format and doi:10.1093/bioinformatics/btv146 for more details
regarding the Cardinal R package. Note that for .imzml files exported from
commercial software, an additional .ibd file may accompany each .imzml,
which should both be located in the same folder.
}
\examples{

# lp1 <- msi_input_param(
#   dp = "pos/",
#   pol = "pos",
#   mz_range = c(300, 1300),
#   mz_res = 5
# )

}
