% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.data.import.R
\name{msi_load_data}
\alias{msi_load_data}
\title{Load .imzml Datasets}
\usage{
msi_load_data(lp)
}
\arguments{
\item{lp}{List containing input parameters from msi_input_param().}
}
\value{
A list of Cardinal MSImagingExperiment objects.
}
\description{
Uses specified input parameters to read files into R using Cardinal.
}
\examples{

# d <- msi_load_data(lp1)

}
