% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{msi_marker_heatmap}
\alias{msi_marker_heatmap}
\title{Top-10 Feature Heatmap}
\usage{
msi_marker_heatmap(so, h_w, h_h, fs_c, fs_r, cl_c, cl_r, rot_c, col1)
}
\arguments{
\item{so}{A Seurat object.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{col1}{Gradient color scheme to use
(must be exactly 4 colors in length).}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a Seurat Object containing
clustered MSI data. By default, the top-10 features per
cluster are returned.
}
\examples{

# p_hmap_top <- msi_marker_heatmap(
#   so = d_seg,
#   h_w = 18,
#   h_h = 24,
#   fs_c = 6,
#   fs_r = 8,
#   cl_c = TRUE,
#   cl_r = TRUE,
#   rot_c = 45,
#   col1 = col_grad()[c(3, 6, 9, 12)]
# )

}
