% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2c.process.peaks.annotation.R
\name{msi_peak_anno}
\alias{msi_peak_anno}
\title{Peak Annotation of MSI Data}
\usage{
msi_peak_anno(ldp, la)
}
\arguments{
\item{ldp}{A data frame of processed peaks.}

\item{la}{A reference m/z list, provided as a data frame.}
}
\value{
A reference-matched data frame containing annotated MSI peaks.
}
\description{
Annotates processed peaks by matching m/z values
from a reference annotation list, such as a complimentary
LC-MS/MS dataset acquired independently from an MSI experiment.
Defaults to matching m/z values between MSI and reference data within 10 mDa.
}
\examples{

# d2a <- msi_peak_anno(
#   ldp = d2,
#   la = an1
# )

}
