% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2a.process.peaks.detection.R
\name{msi_peak_proc}
\alias{msi_peak_proc}
\title{Combined Peak Processing for MSI Datasets}
\usage{
msi_peak_proc(ld, lp, parl, core_perc)
}
\arguments{
\item{ld}{A list of Cardinal MSImagingExperiment objects.}

\item{lp}{A list containing input parameters and file names.}

\item{parl}{Should processing be run in parallel?
Note: only supported by WSL2 or Linux;
defaults to sequential processing on Windows OS.}

\item{core_perc}{Percentage of cores to be used during processing.}
}
\value{
A list containing a data frame of processed peaks and QC metrics.
}
\description{
Performs peak picking and alignment of MSI data files.
}
\examples{
# d2 <- msi_peak_proc(
#   ld = d[["Data.files"]],
#   lp = lp1,
#   parl = TRUE,
#   core_perc = 0.2
# )

}
