% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.chemrich.R
\name{msi_plot_crich}
\alias{msi_plot_crich}
\title{Plot ChemRICH Results}
\usage{
msi_plot_crich(dfcr, cl_name)
}
\arguments{
\item{dfcr}{An individual comparison from a ChemRICH results data frame.}

\item{cl_name}{Variable name to use for assigning classes and grouping
for enrichment analysis.}
}
\value{
A dot plot displaying ChemRICH results,
including cluster size, increased/decreased proportion,
and statistical significance.
}
\description{
Plots ChemRICH results for selected comparison as a dot plot.
}
\examples{

# p_enr <- msi_plot_crich(
#   dfcr = d_enr[d_enr[["Comparison"]] == "comp1", ],
#   cl_name = "label.saturation"
# )

}
