% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3b.normalize.sLOESS.R
\name{msi_plot_tic}
\alias{msi_plot_tic}
\title{Normalization QC}
\usage{
msi_plot_tic(df, var_x, var_y, var_g, y_lim)
}
\arguments{
\item{df}{A data frame containing pixel metadata.}

\item{var_x}{X-axis variable (usually the acquisition order of each pixel).}

\item{var_y}{Y-axis (usually the TIC intensity for each pixel).}

\item{var_g}{Grouping variable (usually the imaging run ID for each sample).}

\item{y_lim}{Upper limit of the y-axis.}
}
\value{
A scatter plot visualizing pixel intensities over time.
}
\description{
Scatter plot visualization to evaluate normalization performance.
}
\examples{

# ptic <- msi_plot_tic(
#   df = d_norm[["pixels"]],
#   var_x = "pixel",
#   var_y = "TIC.norm.tic",
#   var_g = "ID",
#   y_lim = 40000
# )

}
