% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{msi_segment}
\alias{msi_segment}
\title{Seurat-based KNN Clustering and Segmentation}
\usage{
msi_segment(df, sample_no, md, span = 0.1, clus_res = 0.9)
}
\arguments{
\item{df}{A normalized data matrix.}

\item{sample_no}{Sample ID number.}

\item{md}{Number of metadata columns present in data matrix.}

\item{span}{loess span parameter
(proportion of pixels to fit model to;
use lower proportion for tissues with small regions).}

\item{clus_res}{Resolution to use for clustering
(use ?Seurat::FindClusters() documentation for more details;
larger number = more clusters).}
}
\value{
A Seurat object containing segmented clusters for a single sample.
}
\description{
Converts MSI data to a Seurat object and performs spatial
segmentation using the default implementation of KNN
used by Seurat.
}
\examples{

# test <- msi_segment()

}
