% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.stats.univar.R
\name{msi_stat_anova}
\alias{msi_stat_anova}
\title{Multivariate ANOVA}
\usage{
msi_stat_anova(
  matpv = d_seg[["scale.data"]][["data"]][["data.pareto"]],
  matfc = d_seg[["data"]],
  md = d_seg[["pixels"]],
  md_var,
  an = d_seg[["features"]],
  an_name = "Name",
  fc_class1 = FALSE,
  grp_class1 = NULL,
  parl = FALSE,
  core_perc = 0.75
)
}
\arguments{
\item{matpv}{Data matrix for conducting ANOVA.}

\item{matfc}{Data matrix for calculating fold change.}

\item{md}{Metadata from ms_input().}

\item{md_var}{Metadata variables to test by ANOVA.}

\item{an}{Reference annotation data frame.}

\item{an_name}{Variable name containing compound names
(default value is "Name").}

\item{fc_class1}{Logical indicating if compound classes rather
than individual compounds should be tested for significant differences.}

\item{grp_class1}{If fc_class1 is TRUE, which variable should be used
to group individual compounds into separate compound classes?}

\item{parl}{Should stats be run in parallel?}

\item{core_perc}{Not used on Windows; Proportion of cores to use for
conducting statistical analysis in parallel.}
}
\value{
A data frame containing the combined ANOVA and fold change results
for the specified variable(s).
}
\description{
Conducts an ANOVA for selected variables with Tukey's post-hoc
analysis for each compound and combines with fold changes
for each group comparison. The default option is to use
segmented MSI data, stored as a variable named "d_seg."
The defaults can be overridden by specifying additional
function parameters. See ?RegioMSI::msi_data_check for
more details.
}
\examples{

## Specify data
# d_stat <- msi_stat_anova(
#   matpv = d[["data"]][["data.pareto"]],
#   matfc = d[["data"]][["imputed"]],
#   md = d[["data"]][["meta"]],
#   md_var = c("Group", "Cluster"),
#   an = d[["data"]][["anno"]],
#   fc_class1 = FALSE,
#   grp_class1 = "Subclass"
# )

## or use default values
# d_stat <- msi_stat_anova(md_var = c("Group", "Cluster"))

}
