% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.chemrich.R
\name{msi_stat_crich}
\alias{msi_stat_crich}
\title{ChemRICH Analysis Function}
\usage{
msi_stat_crich(
  d_stat,
  d_ref,
  cl_name,
  cmp_name = "Name",
  parl = FALSE,
  core_perc = 0.75
)
}
\arguments{
\item{d_stat}{Univariate results list generated by ms_stat_anova().}

\item{d_ref}{Data frame to use for assigning annotation classes.}

\item{cl_name}{Variable name to use for assigning classes and grouping
for enrichment analysis.}

\item{cmp_name}{(optional) Variable name containing compound names. Default
value is "Name".}

\item{parl}{Should stats be run in parallel?}

\item{core_perc}{Not used on Windows; Proportion of cores to use for
conducting statistical analysis in parallel.}
}
\value{
A ChemRICH results data frame.
}
\description{
Conducts ChemRICH analysis for each comparison in a provided result
generated by msi_stat_anova() (default) or from a single
ChemRICH input (in development). The original enrichment method is adapted
from Chemical Similarity Enrichment Analysis (ChemRICH), which is detailed
in Barupal et. al. 2017 (DOI: 10.1038/s41598-017-15231-w).
Significantly enriched classes are determined by a weighted
Kolmogorov-Smirnov Test adjusted for false discovery rate (FDR).
}
\examples{

# d_enr <- msi_stat_crich(
#   d_stat = d[["stats"]],
#   d_ref = d[["data"]][["anno"]],
#   cl_name = "label.saturation"
# )

}
