% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.stats.univar.R
\name{msi_stat_fc}
\alias{msi_stat_fc}
\title{Group-wise Fold Change Function}
\usage{
msi_stat_fc(
  mat1,
  md,
  md_var,
  an,
  an_name = "Name",
  fc_class = FALSE,
  grp_class = NULL,
  core_perc = 0.75
)
}
\arguments{
\item{mat1}{MSI data frame.}

\item{md}{MSI metadata, formatted as a data frame.}

\item{md_var}{Vector of metadata variables for calculating fold change.}

\item{an}{MSI feature metadata.}

\item{an_name}{Variable name containing compound names
(default value is "Name").}

\item{fc_class}{Logical indicating whether class-wide
fold changes should be calculated. Requires MSI feature metadata if TRUE.}

\item{grp_class}{Name of the class annotation column
to use. Ignored when fc_class is FALSE.}

\item{core_perc}{Not used on Windows; Proportion of cores to use for
conducting statistical analysis in parallel.}
}
\value{
A data frame containing the group-wise
fold changes for each compound.
}
\description{
Calculates the log2 fold changes for all
possible combinations of selected metadata variables.
}
\examples{

## Calculate individual fold changes
# msi_stat_fc(
#   mat1 = d_seg[["data"]],
#   md = d_seg[["pixels"]],
#   md_var = c("Group"),
#   an = d_seg[["features"]]
# )

## Calculate class-based fold changes
# msi_stat_fc(
#   mat1 = d_seg[["data"]],
#   md = d_seg[["pixels"]],
#   md_var = c("Group"),
#   an = d_seg[["features"]],
#   fc_class = TRUE,
#   grp_class = "label.saturation"
# )

}
