% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.stats.univar.R
\name{msi_stat_t}
\alias{msi_stat_t}
\title{Student's T-test of MSI Data}
\usage{
msi_stat_t(df, md, var_g, g1, g2, l_names)
}
\arguments{
\item{df}{Normalized data matrix containing cluster, sample ID,
pixel coordinate, and pixel intensity information.}

\item{md}{Number of metadata columns present in input data.}

\item{var_g}{Treatment group column name.}

\item{g1}{Treatment group 1 name.}

\item{g2}{Treatment group 2 name.}

\item{l_names}{Vector of compound names equal to the number
of compounds present in the given data matrix.}
}
\value{
A data frame containing T-test results for each compound.
}
\description{
Conducts a Student's T-test for each compound.
}
\examples{

# d_stat_t <- msi_stat_t(
#   df = d.p2,
#   md = 2,
#   var_g = "Group",
#   g1 = "F.HO3",
#   g2 = "F.SFA",
#   l_names = an1[["Name"]]
# )

}
