% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{msi_theme1}
\alias{msi_theme1}
\title{Generic Plot Theme}
\usage{
msi_theme1()
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
General plotting theme.
}
\examples{

# msi_theme1()

}
