% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{msi_theme2}
\alias{msi_theme2}
\title{RegioMSI Image Plot Theme}
\usage{
msi_theme2()
}
\value{
ggplot2 theme parameters to replace default image plotting theme.
}
\description{
General image plotting theme.
}
\examples{

# msi_theme2()

}
