% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.segment.clusters.markers.R
\name{msi_umap_panel}
\alias{msi_umap_panel}
\title{MSI UMAP}
\usage{
msi_umap_panel(so, md_list, slot1)
}
\arguments{
\item{so}{A Seurat object.}

\item{md_list}{A vector of character strings indicating
metadata columns for overlaying on a loadings plot.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}
}
\value{
A series of UMAPs with specified metadata overlays.
}
\description{
Generates a panel of UMAPs given a
Seurat object containing PCA and UMAP results.
}
\examples{

# p_umap <- msi_umap_panel(d_integrated,c("col1","col2","col3"),"wnn.umap")

}
