% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{Regio.col.univ}
\alias{Regio.col.univ}
\title{Universal Color Palette}
\usage{
Regio.col.univ()
}
\description{
Combines the npg, aaas, and lancet ggsci palettes for use with datasets
containing up to 29 groups.
}
