% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2c.process.peaks.annotation.R
\name{MSI.data.filter}
\alias{MSI.data.filter}
\title{Peak Annotation of Binned MSI Data}
\usage{
MSI.data.filter(path.rds, an.list.filt)
}
\arguments{
\item{path.rds}{Path to a specific rds file generated by MSI.binpeaks().}

\item{an.list.filt}{A filtered data frame containing annotated MSI peaks from MSI.annotate.peaks().}
}
\value{
A combined Cardinal MSImagingExperiment containing annotated features for all study samples.
}
\description{
Annotates binned MSI data and filters peaks for removing duplicate matched m/z values in complimentary LC-MS/MS data acquired
separately from an MSI experiment.
}
\examples{
list.batch <- list.files()[grepl("processed.b",list.files())]
list.d.anno <- setNames(
  lapply(
    seq.int(
      1,
      length(list.batch),
      1),
    function(x)
      MSI.data.filter(
        list.batch[[x]],
        list.annotated[["Annotated"]]
        )
    ),
  c(list.batch)
  )

## Combine and save processed data
### Combine
if(length(list.batch) > 1) {
  d.processed <- do.call(
    Cardinal::cbind,
    list.d.anno
    )
  }
if(length(list.batch) == 1) {
  d.processed <- list.d.anno[[1]]
  }

### Verify that all samples have been correctly combined
Cardinal::imageData(d.processed)
Cardinal::features(d.processed)
Cardinal::pixelData(d.processed)
Cardinal::run(d.processed)
Cardinal::spectra(
  d.processed,
  "intensity"
  )[,1:25]

### Save data
saveRDS(
  d.processed,
  "data.processed.rds"
  )
gc(reset = T)

}
