% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3a.normalize.format.R
\name{msi_norm_form}
\alias{msi_norm_form}
\title{Normalization Formatting}
\usage{
msi_norm_form(ldp, la, md, ftype)
}
\arguments{
\item{ldp}{A processed MSImagingExperiment.}

\item{la}{A data frame containing reference-matched MSI peaks.}

\item{md}{A data frame containing treatment and group information.}

\item{ftype}{Data to use for normalization (either "annotated" or "all").
"Annotated" filters the MSI data based on features that are matched
to the reference m/z list, whereas "all" retains all features in the
dataset.}
}
\value{
Formatted pixel, feature, and image data for downstream analysis.
}
\description{
Annotates MSImagingExperiment objects by matching m/z values
in a provided reference list. Defaults to matching m/z values
within a 10 mDa threshold.
}
\examples{

# d_norm <- msi_norm_form(
#   ldp = d2a[["Data.filtered"]],
#   la = d2a[["Annotated"]],
#   md = md1,
#   ftype = "annotated"
# )

}
